+++ VWOSdk/Core/VWOClient.cs
@@ -17,6 +17,7 @@
 #pragma warning restore 1587
 
 using System.Collections.Generic;
+using System;
 
 namespace VWOSdk
 {
@@ -371,7 +372,7 @@ namespace VWOSdk
         /// </returns>
         private UserAllocationInfo AllocateVariation(string campaignKey, string userId, BucketedCampaign campaign, Dictionary<string, dynamic> customVariables, Dictionary<string, dynamic> variationTargettingVariable, string apiName = null)
         {
-            Variation TargettedVariation = this.FindTargetedVariation(campaign, campaignKey, userId, customVariables, variationTargettingVariable);
+            Variation TargettedVariation = this.FindTargetedVariation(apiName, campaign, campaignKey, userId, customVariables, variationTargettingVariable);
             if (TargettedVariation != null)
             {
                 return new UserAllocationInfo(TargettedVariation, campaign);
@@ -391,6 +392,7 @@ namespace VWOSdk
                 {
                     if (campaign.Segments.Count > 0)
                     {
+                        string segmentationType = Constants.SegmentationType.PRE_SEGMENTATION;
                         if (customVariables == null)
                         {
                             LogInfoMessage.NoCustomVariables(typeof(IVWOClient).FullName, userId, campaignKey, apiName);
@@ -400,7 +402,8 @@ namespace VWOSdk
                         {
                             variationTargettingVariable = new Dictionary<string, dynamic>();
                         }
-                        if (!this._segmentEvaluator.evaluate(userId, campaignKey, campaign.Segments, customVariables, variationTargettingVariable))
+
+                        if (!this._segmentEvaluator.evaluate(userId, campaignKey, segmentationType, campaign.Segments, customVariables ))
                         {
                             return new UserAllocationInfo();
                         }
@@ -421,44 +424,68 @@ namespace VWOSdk
             return new UserAllocationInfo();
         }
 
-        private Variation FindTargetedVariation(BucketedCampaign campaign, string campaignKey, string userId, Dictionary<string, dynamic> customVariables, Dictionary<string, dynamic> variationTargettingVariable)
+        private Variation FindTargetedVariation(string apiName, BucketedCampaign campaign, string campaignKey, string userId, Dictionary<string, dynamic> customVariables, Dictionary<string, dynamic> variationTargettingVariable)
         {
             if (campaign.IsForcedVariationEnabled)
             {
                 if (variationTargettingVariable == null)
                 {
                     variationTargettingVariable = new Dictionary<string, dynamic>();
-                    variationTargettingVariable.Add("_vwo_user_id", userId);
+                    variationTargettingVariable.Add("_vwoUserId", userId);
                 }
-                List<Variation> whiteListedVariations = this.GetWhiteListedVariationsList(userId, campaign, campaignKey, customVariables, variationTargettingVariable);
+                else
+                {
+                    if (variationTargettingVariable.ContainsKey("_vwoUserId"))
+                    {
+                        variationTargettingVariable["_vwoUserId"] = userId;
+                    }
+                    else
+                    {
+                        variationTargettingVariable.Add("_vwoUserId", userId);
+                    }
+                }
+                List<Variation> whiteListedVariations = this.GetWhiteListedVariationsList(apiName, userId, campaign, campaignKey, customVariables, variationTargettingVariable);
 
-                return this._variationAllocator.TargettedVariation(userId, whiteListedVariations);
+                string status = Constants.WhitelistingStatus.FAILED;
+                string variationString = " ";
+                Variation variation = this._variationAllocator.TargettedVariation(userId, whiteListedVariations);
+                if (variation != null)
+                {
+                    status = Constants.WhitelistingStatus.PASSED;
+                    variationString = $"and variation: {variation.Name} is assigned";
+                }
+                LogInfoMessage.WhitelistingStatus(typeof(IVWOClient).FullName, userId, campaignKey, apiName, variationString, status);
+                return variation;
             }
+            LogInfoMessage.SkippingWhitelisting(typeof(IVWOClient).FullName, userId, campaignKey, apiName);
             return null;
         }
 
-        private List<Variation> GetWhiteListedVariationsList(string userId, BucketedCampaign campaign, string campaignKey, Dictionary<string, dynamic> customVariables, Dictionary<string, dynamic> variationTargettingVariable)
+        private List<Variation> GetWhiteListedVariationsList(string apiName, string userId, BucketedCampaign campaign, string campaignKey, Dictionary<string, dynamic> customVariables, Dictionary<string, dynamic> variationTargettingVariable)
         {
             List<Variation> result = new List<Variation> { };
             foreach (var variation in campaign.Variations.All())
             {
-                bool status;
+                string status = Constants.SegmentationStatus.FAILED;
+                string segmentationType  = Constants.SegmentationType.WHITELISTING;
                 if (variation.Segments.Count == 0)
                 {
-                    status = false;
+                    LogDebugMessage.SkippingSegmentation(typeof(IVWOClient).FullName, userId, campaignKey, apiName, variation.Name);
                 }
                 else
                 {
-                    status = this._segmentEvaluator.evaluate(userId, campaignKey, variation.Segments, customVariables, variationTargettingVariable);
+                    if (this._segmentEvaluator.evaluate(userId, campaignKey, segmentationType, variation.Segments, variationTargettingVariable))
+                    {
+                        status = Constants.SegmentationStatus.PASSED;
+                        result.Add(variation);
+                    }
+                    LogDebugMessage.SegmentationStatus(typeof(IVWOClient).FullName, userId, campaignKey, apiName, variation.Name, status);
                 }
-                if (status)
-                {
-                    result.Add(variation);
-                }
+
             }
             return result;
         }
-        
+
         private UserAllocationInfo GetControlVariation(BucketedCampaign campaign, Variation variation)
         {
             return new UserAllocationInfo(variation, campaign);