+++ VWOSdk/Logger/Messages/LogDebugMessage.cs
@@ -96,5 +96,11 @@ namespace VWOSdk
         public static void ImpressionForPushTag(string file, string properties) {
             Log.Debug($"({file}): impression built for push-tags - {properties}");
         }
+        public static void SkippingSegmentation(string file , string userId, string campaignKey, string apiName, string variationName) {
+            Log.Debug($"({file}): In API: {apiName}, Skipping segmentation for UserId:{userId} in campaing:{campaignKey} for variation: {variationName} as no valid segment is found");
+        }
+        public static void SegmentationStatus(string file , string userId, string campaignKey, string apiName, string variationName, string status) {
+            Log.Debug($"({file}): In API: {apiName}, Whitelisting for UserId:{userId} in campaing:{campaignKey} for variation: {variationName} is: {status}");
+        }
     }
 }